Don't construct ranges from temp containers (gcc v15)
LLVM's/MLIR's `*Ranges` are an immutable "view" into an iterable
container (somewhat analogous to how a `string_view` is a is an immutable
view into a string), meaning they don't actually store the values they
contain, but rather act as a pointer to an actual container somewhere
else.

Hence, if this original container goes out of scope, accessing the
range's values amounts to de-referencing an invalid pointer.

With the version of LLVM being used, constructing a Range from an
initializer list amounts to creating a temporary container, creating a
range from it, then immediately destroying the container. This was
done in a few places in the Dynamatic codebase.

```cpp
// lib/Support/CFG.c:
ValueRange values{selectOp.getTrueValue(), selectOp.getFalseValue()};
```

This updates all instances of this pattern to have the ranges actually
backed by a "physical" `SmallVector`:

```cpp
// lib/Support/CFG.c:
llvm::SmallVector<Value> values{selectOp.getTrueValue(), selectOp.getFalseValue()};
```

Note that I believe that this behavior has been fixed in upstream LLVM,
but we are on too old of an version:

https://github.com/llvm/llvm-project/commit/1297c1125f9c284e0cc0f2bf50d4b7ba519f7309

---

Found while trying to compile Dynamatic with GCC v15. The instance in
`CFG.cpp`, when built with gcc v15 and in release mode, would crash, as
changes in optimization actually caused the memory backing the range to
be dropped quickly enough.