Correctly link SupportBooleanLogic Library (gcc v15)
The `DynamaticExperimentalSupportBooleanLogic` library internally uses
`DynamaticSupportEspresso`, but did not explicitly link against it.

Also, although the `DynamaticExperimentalSupportBooleanLogic` library
was used in some (mostly experimental) dynamatic passes, it was never
actually linked against anywhere:

```bash
rg "DynamaticExperimentalSupportBooleanLogic"
    experimental/lib/Support/BooleanLogic/CMakeLists.txt
    1:add_dynamatic_library(DynamaticExperimentalSupportBooleanLogic
```

This (publically) links it to the ExperimentalSupport lib to ensure all
libraries using it, actually have access to it.

---

Both of these issues became apparent when trying to compile with GCC
v15, likely due to changes in optmisation. For the esecond problem
however, I am not actually sure how this ever worked?!